auth/requirements.txt
# cassandra-driver # as of now, install from source for python 3.12 or greater

pydantic[email]
pydantic-settings
fastapi
uvicorn[standard]

python-jose[cryptography]
passlib[bcrypt]
python-multipart
